# [BugFix] Add an env to disable moe chunking by default