## Purpose
Temp workaround for IMA under high max num batched tokens: https://github.com/vllm-project/vllm/issues/19631
Due to compile incompatibility I feel we shouldn't turn this on by default until we have more comprehensive tests but I'm also open to other options. cc: @zou3519 @bnellnm 

## Test Plan
```
python3 benchmarks/benchmark_latency.py --model $LLAMA_DIR --tensor-parallel-size 8 --trust-remote-code --max_num_batched_tokens 40000
```

## Test Result
Before
```
^[[1;36m(VllmWorker rank=2 pid=3428022)^[[0;0m ERROR 06-13 22:32:19 [multiproc_executor.py:527]   File "/tmp/torchinductor_yeq/y6/cy6oe2twg3y7ikxvey2boducv5ozukmqghzwfjgckbf2c4tatihu.py", line 1282, in call
^[[1;36m(VllmWorker rank=2 pid=3428022)^[[0;0m ERROR 06-13 22:32:19 [multiproc_executor.py:527]     buf52 = empty_strided_cuda(((-32768) + s0, ), (1, ), torch.int32)
^[[1;36m(VllmWorker rank=2 pid=3428022)^[[0;0m ERROR 06-13 22:32:19 [multiproc_executor.py:527]             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
^[[1;36m(VllmWorker rank=2 pid=3428022)^[[0;0m ERROR 06-13 22:32:19 [multiproc_executor.py:527] RuntimeError: Expected result >= 0 to be true, but got false.  (Could this error message be improved?  If so, please report an enhancement request to PyTorch.)
```

After
```
Avg latency: 1.09837638780009 seconds
10% percentile latency: 1.0937033368973061 seconds
25% percentile latency: 1.0950583382800687 seconds
50% percentile latency: 1.0970929424802307 seconds
75% percentile latency: 1.1015077257470693 seconds
90% percentile latency: 1.103359066095436 seconds
99% percentile latency: 1.111307015568018 seconds
```