@yeqcharlotte 2025-06-14
We can turn off inductor `--compilation-config '{"use_inductor": 0}' to achieve similar goal