@bnellnm 2025-06-16
I think we could disable chunking just for cutlass instead. I used the problem sizes from llama4 in the test_cutlass_moe.py unit test and was able to get wrong answers with chunking enabled and no cuda graphs.  So I think there might be an issue with the cutlass moe kernel.